[SDFAB-924] Support P4Runtime translation (#455)
* Support P4Runtime translation test and bug fix

- Remove P4Runtime translation for the switch_info table since we cannot translate the CPU port we are using for that table.
- Add a new job for Jenkins to test P4Runtime translation.
- Move P4Runtime PTF tests to a separate module
- Fix some typos
- Use arch-dependent size port number everywhere in PTF test
- Match ingress and egress port with fabric metadata instead of intrinstic metadata.
- Fix chassis config and port map
  - port 1/0, 2/0, 3/0, and 4/0 are mapped to veth1, 9, 17, and 25.
- Use exact ternary match(with all-one mask) for FabricOptimizedFieldDetectorTest
- Maps SDN port to SDK port when creating expected reports.
- Use SDK port to reset queue report quota register
- Fix ingress port mask for ACL table
- Use unspecified SDN port for packet out test with forwarding flag
- Replace remaining CPU port and recirculation port
- Use new pipeline interpreter API
- Enable p4runtime translation flag by default
- Disable p4runtime translation for bmv2/v1model
- Use long to store instead of integer and use correct numbers in Constants.java
- Use unsigned integer for port number in java code
- Set Stratum incompatible_enable_bfrt_legacy_bytestring_responses to true by default

+++ src/main/java/org/stratumproject/fabric/tna/behaviour/pipeliner/FabricPipeliner.java
@@ -201,7 +201,7 @@ public class FabricPipeliner extends AbstractFabricHandlerBehavior
     protected void initializePipeline() {
         // Set up CPU port for packet-in/out. For packet-out, we support only
         // IPv4 routing when do_forwarding=1.
-        final int cpuPort = capabilities.cpuPort().get();
+        final long cpuPort = capabilities.cpuPort().get();
         flowRuleService.applyFlowRules(
                 egressSwitchInfoRule(cpuPort),
                 ingressVlanRule(cpuPort, false, DEFAULT_VLAN, PORT_TYPE_INTERNAL),
@@ -212,7 +212,7 @@ public class FabricPipeliner extends AbstractFabricHandlerBehavior
 
         // Set up recirculation ports as untagged (used for INT reports and
         // UE-to-UE in UPF pipe).
-        List<Integer> recircPorts = capabilities.isArchTna() ? RECIRC_PORTS : V1MODEL_RECIRC_PORT;
+        List<Long> recircPorts = capabilities.isArchTna() ? RECIRC_PORTS : V1MODEL_RECIRC_PORT;
         recircPorts.forEach(port -> {
             flowRuleService.applyFlowRules(
                     ingressVlanRule(port, false, DEFAULT_VLAN, PORT_TYPE_INTERNAL),
@@ -375,7 +375,7 @@ public class FabricPipeliner extends AbstractFabricHandlerBehavior
         }
     }
 
-    public FlowRule egressSwitchInfoRule(int cpuPort) {
+    public FlowRule egressSwitchInfoRule(long cpuPort) {
         final TrafficTreatment treatment = DefaultTrafficTreatment.builder()
                 .piTableAction(PiAction.builder()
                                        .withId(P4InfoConstants.FABRIC_EGRESS_PKT_IO_EGRESS_SET_SWITCH_INFO)
@@ -419,7 +419,7 @@ public class FabricPipeliner extends AbstractFabricHandlerBehavior
                 .build();
     }
 
-    public FlowRule fwdClassifierRule(int port, Short ethType, short ipEthType, byte fwdType, int priority) {
+    public FlowRule fwdClassifierRule(long port, Short ethType, short ipEthType, byte fwdType, int priority) {
         final TrafficSelector.Builder selectorBuilder = DefaultTrafficSelector.builder()
                 .matchInPort(PortNumber.portNumber(port))
                 .matchPi(PiCriterion.builder()
@@ -446,7 +446,7 @@ public class FabricPipeliner extends AbstractFabricHandlerBehavior
                 .build();
     }
 
-    public FlowRule egressVlanRule(int port, int vlanId, boolean tagged) {
+    public FlowRule egressVlanRule(long port, int vlanId, boolean tagged) {
         final TrafficSelector selector = DefaultTrafficSelector.builder()
                 .add(PiCriterion.builder()
                              .matchExact(P4InfoConstants.HDR_VLAN_ID, vlanId)
